Split signals and events into their own crates
