+++ crates/cli/Cargo.toml
@@ -52,6 +52,14 @@ path = "../project-origins"
 version = "2.1.1"
 path = "../lib"
 
+[dependencies.watchexec-events]
+version = "1.0.0"
+path = "../events"
+
+[dependencies.watchexec-signals]
+version = "1.0.0"
+path = "../signals"
+
 [dependencies.watchexec-filterer-globset]
 version = "1.1.0"
 path = "../filterer/globset"
