+++ crates/events/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## Next (YYYY-MM-DD)
+
+- Split off new `watchexec-events` crate (this one), to have a lightweight library that can parse
+  and generate events and maintain the JSON event format.
