+++ crates/lib/Cargo.toml
@@ -30,6 +30,14 @@ normalize-path = "0.2.0"
 version = "2.0.1"
 features = ["with-tokio"]
 
+[dependencies.watchexec-events]
+version = "1.0.0"
+path = "../events"
+
+[dependencies.watchexec-signals]
+version = "1.0.0"
+path = "../signals"
+
 [dependencies.ignore-files]
 version = "1.1.0"
 path = "../ignore-files"
