+++ crates/lib/README.md
@@ -117,6 +117,9 @@ There are also separate, standalone crates used to build Watchexec which you can
 - **[Command Group](https://docs.rs/command-group)** augments the std and tokio `Command` with
   support for process groups, portable between Unix and Windows.
 
+- **[Event types](https://docs.rs/watchexec-events)** contains the event types used by Watchexec,
+  including the JSON format used for passing event data to child processes.
+
 - **[Ignore files](https://docs.rs/ignore-files)** finds, parses, and interprets ignore files.
 
 - **[Project Origins](https://docs.rs/project-origins)** finds the origin (or root) path of a
