+++ crates/signals/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## Next (YYYY-MM-DD)
+
+- Split off new `watchexec-signals` crate (this one), to have a lightweight library that can parse
+  and represent signals as handled by Watchexec.
