+++ crates/signals/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "watchexec-signals"
+version = "1.0.0"
+
+authors = ["Félix Saparelli <felix@passcod.name>"]
+license = "Apache-2.0 OR MIT"
+description = "Watchexec's signal types"
+keywords = ["watchexec", "signal"]
+
+documentation = "https://docs.rs/watchexec-signals"
+repository = "https://github.com/watchexec/watchexec"
+readme = "README.md"
+
+rust-version = "1.61.0"
+edition = "2021"
+
+[dependencies.miette]
+version = "5.3.0"
+optional = true
+
+[dependencies.thiserror]
+version = "1.0.26"
+optional = true
+
+[target.'cfg(unix)'.dependencies.nix]
+version = "0.26.2"
+features = ["signal"]
+
+[features]
+default = ["parse", "miette"]
+parse = ["dep:thiserror"]
+miette = ["dep:miette"]
